CHAMBER: Move save/load methods to savegame and remove commented code
Move loadGameStream, saveGameStream, syncGameStream and helper functions
from chamber.cpp into savegame.cpp. Remove commented-out original
save/load implementation.
